+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.6.0
 
 - Add support for standard `multipart/form-data` file uploads' request body parsing; remove support for `bodyParserExtendTypes` (swapped [koa-bodyparser](https://www.npmjs.com/package/koa-bodyparser) for [koa-body](https://www.npmjs.com/package/koa-body))
+- Fix: some session related config options always being set to true
 
 ## v1.5.0